Struct neil::Solver

pub struct Solver {
    pub iterations: u64,
    pub initial_temperature: f64,
    pub temperature_reduction: f64,
    pub max_attempts: u64,
    pub max_accepts: u64,
    pub max_rejects: u64,
}

A solver will take a problem and use simulated annealing to try and find an optimal state.

Fields

iterations

The maximum number of iterations to run the algorithm for.

initial_temperature

The initial temperature of the process.

temperature_reduction

The factor to multiply the temperature by each time it is lowered - this should be a number between 0.0 and 1.0.

max_attempts

The maximimum number of attempts to find a new state before lowering the temperature.

max_accepts

The maximum number of accepted new states before lowering the temperature.

max_rejects

The maximum number of rejected states before terminating the process.

Methods

impl Solver

fn new() -> Solver

Construct the new default solver.

fn build_new<F>(builder: F) -> Solver where F: FnOnce(&mut Solver)

Construct a new solver with a given builder function.

fn solve<P>(&self, problem: &P) -> P::State where P: Problem

Run the solver on the given problem.